Initial checkin
This is just checking in C soruce files. It is a bit hard to unravel assets in the project because some
are from packages that I downloaded (Really wishing these game engines had a good module system to keep modules
separate and to specify dependencies). Anway this is to make the code visible for people who are watching
a few youtube videos that I made.